Add files via upload
In this repository there is a webpage having some information of different companies which are connected through linking and there is an application form along with each company i.e. in any company you may apply and after fillng the aplication form there is a thanyou message displayed on your screen as well!